requirements.txt
alabaster==0.7.12
arabic-reshaper==2.1.3
asgiref==3.5.2
astatine==0.3.2
astor==0.8.1
astpretty==3.0.0
astroid==2.11.7
asttokens==2.0.8
attrs==22.1.0
autoflake==1.4
Babel==2.10.3
backports.zoneinfo==0.2.1
bandit==1.7.4
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
certifi
charset-normalizer==2.1.1
click==8.1.3
cognitive-complexity==1.3.0
coverage==6.4.4
darglint==1.8.1
dill==0.3.5.1
distlib==0.3.5
Django==4.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-datetime-widget==0.9.3
django-environ==0.9.0
django-filter==22.1
django-multiselectfield==0.1.12
django-summernote==0.8.20.0
django-utils-six==2.0
djangorestframework==3.13.1
dlint==0.13.0
doc8==0.11.2
docformatter==1.5.0
docutils==0.17.1
domdf-python-tools==3.3.0
environ==1.0
eradicate==2.1.0
exceptiongroup==1.0.0rc8
filelock==3.8.0
flake8==4.0.1
flake8-2020==1.6.1
flake8-aaa==0.12.2
flake8-annotations==2.9.1
flake8-annotations-complexity==0.0.7
flake8-annotations-coverage==0.0.6
flake8-bandit==3.0.0
flake8-black==0.3.3
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.4.0
flake8-bugbear==22.7.1
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-django==1.1.5
flake8-docstrings==1.6.0
flake8-encodings==0.5.0.post1
flake8-eradicate==1.3.0
flake8-executable==2.1.1
flake8-expression-complexity==0.0.11
flake8-fixme==1.1.1
flake8-functions==0.0.7
flake8-functions-names==0.3.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.1
flake8-isort==4.2.0
flake8-literal==1.2.6
flake8-logging-format==0.7.5
flake8-markdown==0.3.0
flake8-mutable==1.2.0
flake8-no-pep420==2.3.0
flake8-noqa==1.2.9
flake8-pie==0.16.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-pyi==22.8.1
flake8-pylint==0.1.3
flake8-pytest-style==1.6.0
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-secure-coding-standard==1.3.0
flake8-simplify==0.19.3
flake8-slots==0.1.5
flake8-string-format==0.3.0
flake8-tidy-imports==4.8.0
flake8-typing-imports==1.12.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.3
flake8-variables-names==0.0.5
flake8-warnings==0.2.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
html5lib==1.1
hypothesis==6.54.3
hypothesmith==0.1.9
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
Jinja2==3.1.2
jupyterlab-flake8==0.7.1
lark-parser==0.12.0
lazy-object-proxy==1.7.1
libcst==0.4.7
MarkupSafe==2.1.1
mccabe==0.6.1
mr-proper==0.0.7
mypy-extensions==0.4.3
mysqlclient==2.1.1
natsort==8.1.0
packaging==21.3
pandas-vet==0.2.3
pathspec==0.9.0
pbr==5.10.0
pep8-naming==0.12.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pybetter==0.4.1
pycln==1.3.5
pycodestyle==2.8.0
pydocstyle==6.1.1
pyemojify==0.2.0
pyflakes==2.4.0
Pygments==2.13.0
pylint==2.14.5
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-sugar==0.9.5
python-dateutil==2.8.2
python-dev-tools==2022.5.27
pytz==2022.2.1
pyupgrade==2.37.3
PyYAML==6.0
removestar==1.3.1
requests==2.28.1
restructuredtext-lint==1.4.0
six==1.16.0
slugify==0.0.1
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
ssort==0.11.6
stdlib-list==0.8.0
stevedore==4.0.0
termcolor==1.1.0
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tox==3.25.1
tox-travis==0.12
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.3.0
untokenize==0.1.1
urllib3==1.26.11
virtualenv==20.16.3
webencodings==0.5.1
wemake-python-styleguide==0.16.1
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.8.1
